Added defined_vars option for combined adjoint
Added functionality for a defined_vars input to the combined adjoint
test that allows the user to provide specified values for any/all
variables within the analysis stack. This makes use of the global
naming convention, which was also updated slightly for ease of use
prior to performing an analysis.