Added pre-commit check for formatting/unittest
The pre-commit will locally check for black formatting and ensure that
unittests all pass before changes are committed to the repository. If
the black formatting hook fails, it will automatically reformat the
files, but the user must add/stage the modified files again to reflect
the additional updates from the auto-formatting.