Updated logic for analysis initialization
This fixed a bug that was related to resetting the 'analyzed' flag for
analysis streams that diverge and reconverge. Previously, this check was
only using information from the previous analysis. Now, the check uses
information about all sub-analyses for an object. The multi-objective
example was updated with additional analyses to reflect a System
structure that previously failed derivative checks, but now passes with
the updated logic.